Reorganize validators
