+++ domains/schema/transforms/urlTransform.ts
@@ -1,0 +1,21 @@
+import { z } from 'zod'
+
+/**
+ * Validate url
+ *
+ * @param value
+ */
+export const urlTransform = async (value: string, ctx: z.RefinementCtx) => {
+  const { formatMessage } = useIntl()
+
+  try {
+    new URL(value)
+    return encodeURI(value)
+  } catch (error) {
+    ctx.addIssue({
+      code: z.ZodIssueCode.custom,
+      message: formatMessage('errors_invalid_url'),
+    })
+    return z.NEVER
+  }
+}
